Re-org: Video.cache() + Thumbnail.cache()
